[CHORE] ViewModel과 Controller 분리
